+++ packages/ecommerce-app-base/.storybook/main.ts
@@ -5,7 +5,6 @@ const config: StorybookConfig = {
   addons: [
     '@storybook/addon-links',
     '@storybook/addon-essentials',
-    '@storybook/addon-onboarding',
     '@storybook/addon-interactions',
     '@storybook/addon-designs',
     '@storybook/addon-a11y',