+++ packages/ecommerce-app-base/src/Editor/SortableComponent.stories.tsx
@@ -1,15 +1,17 @@
+import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import { productsList } from '../__mocks__';
-import { SortableList } from './SortableList';
 import { integration } from '../__mocks__/storybook/integration';
 import { IntegrationProvider } from './IntegrationContext';
 import { Integration } from '../types';
 import { decorators } from '../__mocks__/storybook/decorators';
+import { SortableComponent } from './SortableComponent';
+import { sdk } from '../__mocks__/storybook/sdk';
 
-type Args = typeof SortableList & Pick<Integration, 'productCardVersion'>;
+type Args = typeof SortableComponent & Pick<Integration, 'productCardVersion'>;
 
 const meta: Meta<Args> = {
-  component: SortableList,
+  component: SortableComponent,
   tags: ['autodocs'],
   argTypes: {
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -25,14 +27,25 @@ const meta: Meta<Args> = {
 
 export default meta;
 
-type Story = StoryObj<typeof SortableList>;
+type Story = StoryObj<typeof SortableComponent>;
 
 export const Default: Story = {
   args: {
-    skuType: 'Product',
-    productPreviews: productsList,
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    sdk: sdk,
     disabled: false,
-    deleteFn: () => {},
+    onChange: (newSkus: string[]) => {
+      // it'd be a nice to have, if this would actually update productsList and the UI, but storybook
+      // does not rerender unless args are directly changed, and I can't find a workaround in a reasonable
+      // amount of time.  So just console logging for now.
+      console.log('[ SortableComponent.stories.tsx ] onChange() new SKUs:', newSkus);
+    },
+    config: {},
+    skus: productsList.map((p) => p.sku),
+    fetchProductPreviews: async () => {
+      return productsList;
+    },
   },
   render: (args) => {
     return (
@@ -40,7 +53,7 @@ export const Default: Story = {
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         integration={integration({ productCardVersion: args.productCardVersion })}>
-        <SortableList {...args} />
+        <SortableComponent {...args} />
       </IntegrationProvider>
     );
   },