+++ packages/ecommerce-app-base/src/__mocks__/storybook/integration.tsx
@@ -1,7 +1,5 @@
 import { Integration } from '../../types';
 import { productsList } from '../products';
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
 import logo from '../logo.svg';
 
 export function integration(